Switch glob importing strategy to avoid error
Switch from using a rollup plugin to a babel plugin

This resolves an odd error I was getting when saving JS files.

We may still want to think more about how the toolkit imports files.

Maybe we leave it like this for now but expect end users are likely to modify it?